pyproject.toml
[tool.poetry]
name = "coava-backend"
version = "0.1.0"
description = ""
authors = ["puilin <suryblue@naver.com>"]
license = "MIT License"
readme = "README.md"
packages = [{include = "coava_backend"}]

[tool.poetry.dependencies]
python = "^3.9"
torch = [
    {url = "https://download.pytorch.org/whl/cpu/torch-2.0.0-cp311-none-macosx_11_0_arm64.whl", markers = "sys_platform == 'darwin' and platform_machine == 'arm64'"},
    {version = "2.0.0", markers = "sys_platform == 'Windows'"},
    {url = "https://download.pytorch.org/whl/cpu/torch-2.0.0-cp311-none-macosx_10_9_x86_64.whl", markers = "sys_platform == 'darwin' and platform_machine == 'x86_64'"}
]
django = "^4.1.7"
djangorestframework = "^3.14.0"
mysqlclient = "^2.1.1"
django-dotenv = "^1.4.2"
argon2-cffi = "^21.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
